backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agents-platform-backend"
version = "0.1.0"
description = "Agent generation and deployment platform backend - Azure AI Foundry"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "azure-identity>=1.16.0",
    "azure-ai-projects>=1.0.0b4",
    "azure-ai-inference>=1.0.0b6",
    "azure-ai-evaluation>=1.1.0",
    "azure-cosmos>=4.7.0",
    "azure-ai-contentsafety>=1.0.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "ruff>=0.5.0",
    "mypy>=1.10",
]
